+++ internal/extension_repo/goja_plugin_types/app.d.ts
@@ -723,7 +723,8 @@ declare namespace $app {
      * @file internal/continuity/hook_events.go
      * @description
      * WatchHistoryItemRequestedEvent is triggered when a watch history item is requested.
-     * Prevent default to skip getting the watch history item from the file cache, in this case the event should have a valid WatchHistoryItem object or set it to nil to indicate that the watch history item was not found.
+     * Prevent default to skip getting the watch history item from the file cache, in this case the event should have a valid WatchHistoryItem object
+     *     or set it to nil to indicate that the watch history item was not found.
      */
     function onWatchHistoryItemRequested(cb: (event: WatchHistoryItemRequestedEvent) => void): void;
 
@@ -870,10 +871,10 @@ declare namespace $app {
      * @event DiscordPresenceAnimeActivityRequestedEvent
      * @file internal/discordrpc/presence/hook_events.go
      * @description
-     * DiscordPresenceAnimeActivityRequestedEvent is triggered when anime activity is requested, after the [animeActivity] is processed, and right before the activity is sent to queue.
-     * There is no guarantee as to when or if the activity will be successfully sent to discord.
-     * Note that this event is triggered every 6 seconds or so, avoid heavy processing or perform it only when the activity is changed.
-     * Prevent default to stop the activity from being sent to discord.
+     * DiscordPresenceAnimeActivityRequestedEvent is triggered when anime activity is requested, after the [animeActivity] is processed, and right
+     *     before the activity is sent to queue. There is no guarantee as to when or if the activity will be successfully sent to discord. Note that
+     *     this event is triggered every 6 seconds or so, avoid heavy processing or perform it only when the activity is changed. Prevent default to
+     *     stop the activity from being sent to discord.
      */
     function onDiscordPresenceAnimeActivityRequested(cb: (event: DiscordPresenceAnimeActivityRequestedEvent) => void): void;
 
@@ -901,10 +902,10 @@ declare namespace $app {
      * @event DiscordPresenceMangaActivityRequestedEvent
      * @file internal/discordrpc/presence/hook_events.go
      * @description
-     * DiscordPresenceMangaActivityRequestedEvent is triggered when manga activity is requested, after the [mangaActivity] is processed, and right before the activity is sent to queue.
-     * There is no guarantee as to when or if the activity will be successfully sent to discord.
-     * Note that this event is triggered every 6 seconds or so, avoid heavy processing or perform it only when the activity is changed.
-     * Prevent default to stop the activity from being sent to discord.
+     * DiscordPresenceMangaActivityRequestedEvent is triggered when manga activity is requested, after the [mangaActivity] is processed, and right
+     *     before the activity is sent to queue. There is no guarantee as to when or if the activity will be successfully sent to discord. Note that
+     *     this event is triggered every 6 seconds or so, avoid heavy processing or perform it only when the activity is changed. Prevent default to
+     *     stop the activity from being sent to discord.
      */
     function onDiscordPresenceMangaActivityRequested(cb: (event: DiscordPresenceMangaActivityRequestedEvent) => void): void;
 
@@ -1221,9 +1222,9 @@ declare namespace $app {
      * @file internal/api/metadata/hook_events.go
      * @description
      * AnimeEpisodeMetadataEvent is triggered when anime episode metadata is available and is about to be returned.
-     * In the current implementation, episode metadata is requested for display purposes. It is used to get a more complete metadata object since the original AnimeMetadata object is not complete.
-     * This event is triggered after [AnimeEpisodeMetadataRequestedEvent].
-     * If the modified episode metadata is nil, an empty EpisodeMetadata object will be returned.
+     * In the current implementation, episode metadata is requested for display purposes. It is used to get a more complete metadata object since the
+     *     original AnimeMetadata object is not complete. This event is triggered after [AnimeEpisodeMetadataRequestedEvent]. If the modified episode
+     *     metadata is nil, an empty EpisodeMetadata object will be returned.
      */
     function onAnimeEpisodeMetadata(cb: (event: AnimeEpisodeMetadataEvent) => void): void;
 
@@ -1301,8 +1302,8 @@ declare namespace $app {
      * PlaybackLocalFileDetailsRequestedEvent is triggered when the local files details for a specific path are requested.
      * This event is triggered right after the media player loads an episode.
      * The playback manager uses the local files details to track the progress, propose next episodes, etc.
-     * In the current implementation, the details are fetched by selecting the local file from the database and making requests to retrieve the media and anime list entry.
-     * Prevent default to skip the default fetching and override the details.
+     * In the current implementation, the details are fetched by selecting the local file from the database and making requests to retrieve the media
+     *     and anime list entry. Prevent default to skip the default fetching and override the details.
      */
     function onPlaybackLocalFileDetailsRequested(cb: (event: PlaybackLocalFileDetailsRequestedEvent) => void): void;
 
@@ -1324,7 +1325,8 @@ declare namespace $app {
      * @description
      * PlaybackStreamDetailsRequestedEvent is triggered when the stream details are requested.
      * Prevent default to skip the default fetching and override the details.
-     * In the current implementation, the details are fetched by selecting the anime from the anime collection. If nothing is found, the stream is still tracked.
+     * In the current implementation, the details are fetched by selecting the anime from the anime collection. If nothing is found, the stream is
+     *     still tracked.
      */
     function onPlaybackStreamDetailsRequested(cb: (event: PlaybackStreamDetailsRequestedEvent) => void): void;
 
@@ -2724,6 +2726,7 @@ declare namespace $app {
         anidbId: number;
         currentEpisodeCount: number;
         _isNakamaEntry: boolean;
+        nakamaLibraryData?: Anime_NakamaEntryLibraryData;
     }
 
     /**
@@ -2867,6 +2870,10 @@ declare namespace $app {
          */
         libraryData?: Anime_EntryLibraryData;
         /**
+         * Library data from Nakama
+         */
+        nakamaLibraryData?: Anime_NakamaEntryLibraryData;
+        /**
          * AniList list data
          */
         listData?: Anime_EntryListData;
@@ -2955,6 +2962,14 @@ declare namespace $app {
     interface Anime_MissingEpisodes {
         episodes?: Array<Anime_Episode>;
         silencedEpisodes?: Array<Anime_Episode>;
+    }
+
+    /**
+     * - Filepath: internal/library/anime/entry_library_data.go
+     */
+    interface Anime_NakamaEntryLibraryData {
+        unwatchedCount: number;
+        mainFileCount: number;
     }
 
     /**