+++ internal/library/anime/entry.go
@@ -29,7 +29,8 @@ type (
 		AnidbId             int                `json:"anidbId"`
 		CurrentEpisodeCount int                `json:"currentEpisodeCount"`
 
-		IsNakamaEntry bool `json:"_isNakamaEntry"`
+		IsNakamaEntry     bool                    `json:"_isNakamaEntry"`
+		NakamaLibraryData *NakamaEntryLibraryData `json:"nakamaLibraryData,omitempty"`
 	}
 
 	// EntryListData holds the details of the AniList entry.