+++ internal/torrentstream/collection.go
@@ -47,6 +47,21 @@ func (r *Repository) HydrateStreamCollection(opts *HydrateStreamCollectionOption
 		}
 		if *list.Status == anilist.MediaListStatusCurrent {
 			currentlyWatching = list
+			continue
+		}
+	}
+	for _, list := range lists {
+		if list.Status == nil {
+			continue
+		}
+		if *list.Status == anilist.MediaListStatusRepeating {
+			if currentlyWatching == nil {
+				currentlyWatching = list
+			} else {
+				for _, entry := range list.Entries {
+					currentlyWatching.Entries = append(currentlyWatching.Entries, entry)
+				}
+			}
 			break
 		}
 	}